pyproject.toml
[project]
name = "mcp-email-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.0",
    "python-dotenv>=1.1.0",
]
